requirements.txt
opencv-python-headless == 4.8.0.76
robotpy-wpimath == 2023.4.3.1
pyntcore == 2023.4.3.0
Pillow == 10.0.1
